# apps/gha-badge-generator/src/main/resources/openapi/openapi.yaml
openapi: "3.0.0"
info:
  description: "A reactive badge generator for GitHub Actions workflows"
  version: "1.0.0"
  title: "GitHub Action Badge Generator"
  contact:
    name: "Domenico Sibilio"
    email: "domenicosibilio@gmail.com"
tags: 
  - name: Badges API
paths:
  /badges/{message}:
    get:
      summary: "Generate a static badge"
      operationId: "getBadge"
      parameters: 
        - in: path
          name: message
          required: true
          schema:
            type: string
            minLength: 1
        - in: query
          name: label
          schema:
            type: string
        - in: query
          name: messageColor
          schema:
            $ref: "#/components/schemas/NamedColor"
        - in: query
          name: labelColor
          schema:
            $ref: "#/components/schemas/NamedColor"
        - in: query
          name: logo
          schema:
            type: string
      responses:
        200:
          description: "The request has been accepted"
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
        400:
          description: "Message validation exception"
        500:
          description: "Internal server error"
      tags:
        - Badges API
  /badges/github/{owner}/{repository}/{workflowId}:
    get:
      summary: "Generate a dynamic GitHub Action workflow badge"
      operationId: "getGitHubWorkflowBadge"
      parameters: 
        - in: path
          name: owner
          required: true
          schema:
            type: string
            minLength: 1
        - in: path
          name: repository
          required: true
          schema:
            type: string
            minLength: 1
        - in: path
          name: workflowId
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        200:
          description: "The request has been accepted"
          content:
            image/svg+xml:
              schema:
                type: string
                format: binary
        400:
          description: "Message validation exception"
        500:
          description: "Internal server error"
      tags:
        - Badges API
components:
  schemas:
    NamedColor:
      type: string
      enum: [ BRIGHTGREEN, GREEN, YELLOW, YELLOWGREEN, ORANGE, RED, BLUE, GREY, LIGHTGREY ]